requirements.txt
# ─── Apache Airflow ──────────────────────────────────────────────────────────
apache-airflow==2.9.0
apache-airflow-providers-amazon==8.20.0

# ─── Apache Spark ────────────────────────────────────────────────────────────
pyspark==3.5.0

# ─── Elasticsearch ───────────────────────────────────────────────────────────
elasticsearch==8.12.0
elastic-transport==8.13.1

# ─── Data / Finance ──────────────────────────────────────────────────────────
yfinance==0.2.37
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.0
xgboost==3.2.0

# ─── AWS / S3 ────────────────────────────────────────────────────────────────
boto3==1.34.29
botocore==1.34.29
awscli-local==0.22.0

# ─── HTTP / Twitch ───────────────────────────────────────────────────────────
requests==2.31.0

# ─── Config ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
PyYAML==6.0.1

# ─── Utilities ───────────────────────────────────────────────────────────────
pyarrow==14.0.2
fastparquet==2024.2.0
